Separated the individual Model Provider logic from the SpringBoot Controllers
