+++ java/src/main/java/com/datasqrl/ai/models/bedrock/BedrockChatModel.java
@@ -13,7 +13,7 @@ public enum BedrockChatModel implements GenericLanguageModel {
 //  Until then, we can use the llama3-7b tokenizer as a temporary solution, as the tokenizer is currently only used to
 //  prevent context window overflow
   LLAMA3_70B("meta.llama3-70b-instruct-v1:0", "meta-llama/Meta-Llama-3-8B-Instruct", 8192, 1024),
-  LLAMA3_7B("meta.llama3-8b-instruct-v1:0", "meta-llama/Meta-Llama-3-8B-Instruct", 8192, 512);
+  LLAMA3_8B("meta.llama3-8b-instruct-v1:0", "meta-llama/Meta-Llama-3-8B-Instruct", 8192, 512);
 
   final String modelName;
   final String tokenizerName;
